# backend/src/statsController.ts
import { Controller, Delete, Get, Query, Route, Security } from '@tsoa/runtime';

import { IPlayerStats, IMatchStats, IMatchMapStats } from '../../common';

import * as MatchService from './matchService';
import * as StatsLogger from './statsLogger';

@Route('/api/stats')
@Security('bearer_token_optional')
export class StatsController extends Controller {
	/**
	 * Get global player statistics.
	 */
	@Get('/players')
	async getPlayersStats(): Promise<IPlayerStats[]> {
		return StatsLogger.getPlayersStats();
	}

	/**
	 * Get the stats of all the players for a match.
	 */
	@Get('/players/match')
	async getMatchPlayersStats(@Query('id') id: string): Promise<IPlayerStats[]> {
		return StatsLogger.getMatchPlayersStats(id);
	}

	/**
	 * Get global match statistics.
	 */
	@Get('/matches')
	async getMatchesStats(): Promise<IMatchStats[]> {
		return StatsLogger.getMatchesStats();
	}

	/**
	 * Get match stats for a specific match.
	 */
	@Get('/match')
	async getMatchStats(@Query('id') id: string): Promise<IMatchStats> {
		return StatsLogger.getMatchStats(id);
	}

	/**
	 * Get match map stats for a specific match map.
	 */
	@Get('/match/map')
	async getMatchMapStats(
		@Query('id') id: string,
		@Query('map') map: string
	): Promise<IMatchMapStats> {
		return StatsLogger.getMatchMapStats(id, map);
	}

	/**
	 * Get the total stats of a player.
	 */
	@Get('/player')
	async getPlayerStats(@Query('id') id: string): Promise<IPlayerStats> {
		return StatsLogger.getPlayerStats(id);
	}

	/**
	 * Get the stats of a player, divided by match.
	 */
	@Get('/matches/player')
	async getPlayerMatchesStats(@Query('id') id: string): Promise<IPlayerStats[]> {
		return StatsLogger.getPlayerMatchesStats(id);
	}

	/**
	 * Get all the players for a team.
	 */
	@Get('/team')
	async getTeamPlayers(@Query('id') id: string): Promise<string[]> {
		return StatsLogger.getTeamPlayers(id);
	}

	@Delete()
	async deleteStats() {
		if (MatchService.getAllLive().length > 0) {
			throw { status: 409, message: `Can't modify database while matches are running.` };
		}

		StatsLogger.deleteAllStats();
	}
}