# backend/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IPlayerStats": {
				"properties": {
					"steamId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"kills": {
						"type": "number",
						"format": "double"
					},
					"deaths": {
						"type": "number",
						"format": "double"
					},
					"assists": {
						"type": "number",
						"format": "double"
					},
					"hits": {
						"type": "number",
						"format": "double"
					},
					"headshots": {
						"type": "number",
						"format": "double"
					},
					"rounds": {
						"type": "number",
						"format": "double"
					},
					"damages": {
						"type": "number",
						"format": "double"
					},
					"kd": {
						"type": "number",
						"format": "double"
					},
					"hsPct": {
						"type": "number",
						"format": "double"
					},
					"adr": {
						"type": "number",
						"format": "double"
					},
					"map": {
						"type": "string"
					},
					"matchId": {
						"type": "string"
					}
				},
				"required": [
					"steamId",
					"name",
					"kills",
					"deaths",
					"assists",
					"hits",
					"headshots",
					"rounds",
					"damages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMatchStats": {
				"properties": {
					"matchId": {
						"type": "string"
					},
					"teamA": {
						"type": "string"
					},
					"teamAScore": {
						"type": "number",
						"format": "double"
					},
					"teamB": {
						"type": "string"
					},
					"teamBScore": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["matchId", "teamA", "teamAScore", "teamB", "teamBScore", "timestamp"],
				"type": "object",
				"additionalProperties": false
			},
			"IMatchMapStats": {
				"properties": {
					"matchId": {
						"type": "string"
					},
					"map": {
						"type": "string"
					},
					"teamA": {
						"type": "string"
					},
					"teamAScore": {
						"type": "number",
						"format": "double"
					},
					"teamB": {
						"type": "string"
					},
					"teamBScore": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"matchId",
					"map",
					"teamA",
					"teamAScore",
					"teamB",
					"teamBScore",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TMatchState": {
				"type": "string",
				"enum": ["ELECTION", "MATCH_MAP", "FINISHED"],
				"description": "Possible match states."
			},
			"ITeam": {
				"description": "Team.",
				"properties": {
					"passthrough": {
						"type": "string",
						"description": "Passthrough data to identify team in other systems.\nWill be present in every response/webhook."
					},
					"name": {
						"type": "string",
						"description": "Team name."
					},
					"advantage": {
						"type": "number",
						"format": "double",
						"description": "Advantage in map wins, useful for double elimination tournament finals."
					},
					"playerSteamIds64": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Steam ids of players in \"Steam ID 64\" format. Will be forced into this team."
					}
				},
				"required": ["name", "advantage"],
				"type": "object",
				"additionalProperties": false
			},
...
			"IElectionStepSkip": {
				"description": "Election steps which will remove a map from the map pool.",
				"properties": {
					"map": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/IRandomMapBan"
							},
							{
								"$ref": "#/components/schemas/IBanMap"
							}
						]
					}
				},
				"required": ["map"],
				"type": "object",
				"additionalProperties": false
			},
			"IElectionStep": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IElectionStepAdd"
					},
					{
						"$ref": "#/components/schemas/IElectionStepSkip"
					}
				],
				"description": "Election step to either add map to the list of match maps. Or to remove a map from the map pool."
			},
...
			"IElection": {
				"properties": {
					"state": {
						"$ref": "#/components/schemas/TElectionState"
					},
					"teamX": {
						"$ref": "#/components/schemas/TTeamAB"
					},
					"teamY": {
						"$ref": "#/components/schemas/TTeamAB"
					},
					"remainingMaps": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Will be the same as the mapPool from the match, but will shrink when maps get picked, banned or randomly chosen."
					},
					"currentStep": {
						"type": "number",
						"format": "double",
						"description": "Index of the current electionSteps of the match."
					},
					"currentSubStep": {
						"$ref": "#/components/schemas/TStep",
						"description": "Toggles between MAP and SIDE"
					},
					"currentStepMap": {
						"type": "string",
						"description": "Current set map of the current selection step."
					},
					"currentAgree": {
						"properties": {
							"teamB": {
								"type": "string",
								"nullable": true
							},
							"teamA": {
								"type": "string",
								"nullable": true
							}
						},
						"required": ["teamB", "teamA"],
						"type": "object",
						"description": "Holds the wanted maps of each team."
					},
					"currentRestart": {
						"properties": {
							"teamB": {
								"type": "boolean"
							},
							"teamA": {
								"type": "boolean"
							}
						},
						"required": ["teamB", "teamA"],
						"type": "object",
						"description": "The election process can be restarted if both teams vote for it."
					}
				},
				"required": [
					"state",
					"remainingMaps",
					"currentStep",
					"currentSubStep",
					"currentAgree",
					"currentRestart"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGameServer": {
				"properties": {
					"ip": {
						"type": "string"
					},
					"port": {
						"type": "number",
						"format": "double"
					},
					"rconPassword": {
						"type": "string"
					},
					"hideRconPassword": {
						"type": "boolean",
						"description": "If plebs (client without an admin token) create a match the hideRconPassword attribute is set to true.\nThis will prevent executing rcon commands from the frontend by the (unauthorized) user."
					}
				},
				"required": ["ip", "port", "rconPassword"],
				"type": "object",
				"additionalProperties": false
			},
...
			"IMatchMap": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Map name, e.g. de_anubis."
					},
					"knifeForSide": {
						"type": "boolean"
					},
					"startAsCtTeam": {
						"$ref": "#/components/schemas/TTeamAB",
						"description": "may change after knife round"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchMapSate"
					},
					"knifeWinner": {
						"$ref": "#/components/schemas/TTeamAB",
						"description": "Winner of the knife round which is able to or already has picked a starting side."
					},
					"readyTeams": {
						"properties": {
							"teamB": {
								"type": "boolean"
							},
							"teamA": {
								"type": "boolean"
							}
						},
						"required": ["teamB", "teamA"],
						"type": "object"
					},
					"knifeRestart": {
						"properties": {
							"teamB": {
								"type": "boolean"
							},
							"teamA": {
								"type": "boolean"
							}
						},
						"required": ["teamB", "teamA"],
						"type": "object"
					},
					"score": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": ["teamB", "teamA"],
						"type": "object",
						"description": "Current score of both teams."
					},
					"overTimeEnabled": {
						"type": "boolean",
						"description": "If overtime is enabled (mp_overtime_enable)."
					},
					"overTimeMaxRounds": {
						"type": "number",
						"format": "double",
						"description": "Max rounds in overtime (mp_overtime_maxrounds)."
					},
					"maxRounds": {
						"type": "number",
						"format": "double",
						"description": "Max rounds (mp_maxrounds)."
					}
				},
				"required": [
					"name",
					"knifeForSide",
					"startAsCtTeam",
					"state",
					"readyTeams",
					"knifeRestart",
					"score",
					"overTimeEnabled",
					"overTimeMaxRounds",
					"maxRounds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TMatchEndAction": {
				"type": "string",
				"enum": ["KICK_ALL", "QUIT_SERVER", "NONE"]
			},
...
			"TLogUnion": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ILogChat"
					},
					{
						"$ref": "#/components/schemas/ILogSystem"
					}
				]
			},
...
			"IPlayer": {
				"description": "Player.",
				"properties": {
					"steamId64": {
						"type": "string",
						"description": "Steam ID 64"
					},
					"name": {
						"type": "string",
						"description": "Name."
					},
					"team": {
						"$ref": "#/components/schemas/TTeamAB",
						"description": "Current team as they joined with `.team`.\nIf the player's steam id is in the team's `playerSteamIds64`\nthis cannot be changed and is always set to the team."
					},
					"side": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TTeamSides"
							}
						],
						"nullable": true,
						"description": "Current ingame side."
					},
					"online": {
						"type": "boolean",
						"description": "Player currently on the game server (online)?"
					}
				},
				"required": ["steamId64", "name"],
				"type": "object",
				"additionalProperties": false
			},
			"TMatchMode": {
				"type": "string",
				"enum": ["SINGLE", "LOOP"],
				"description": "Possible match modes."
			},
			"IMatch": {
				"properties": {
					"id": {
						"type": "string",
						"description": "tmt2 identifier for this match"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchState"
					},
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool.\nWorkshop maps are possible (contain numbers only).\nFriendly names for players when picking/banning maps can be added after a \"/\" delimiter and is advised for workshop maps.\n\nExample:\n```\n[\n   \"de_ancient\",\n   \"de_anubis/anubis\",\n   \"de_inferno\",\n   \"de_mirage\",\n   \"de_nuke\",\n   \"de_overpass\",\n   \"de_vertigo\",\n   \"3070923343/fy_pool_day\"\n]\n```"
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team A for this match.\nTeam A will always be Team A in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team B for this match.\nTeam B will always be Team B in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"electionSteps": {
						"items": {
							"$ref": "#/components/schemas/IElectionStep"
						},
						"type": "array",
						"description": "List of election steps to determine the played map(s)."
					},
					"election": {
						"$ref": "#/components/schemas/IElection",
						"description": "Data for the election process."
					},
					"gameServer": {
						"$ref": "#/components/schemas/IGameServer"
					},
					"logSecret": {
						"type": "string",
						"description": "Log secret that is given as part of the url to the CS2 server as a log receiver (logaddress_add_http)."
					},
					"parseIncomingLogs": {
						"type": "boolean",
						"description": "Indicates if incoming logs from the CS2 server are parsed (otherwise they will be dropped without any action).\nWill be set to true if match is loaded from storage (after a short delay)."
					},
					"matchMaps": {
						"items": {
							"$ref": "#/components/schemas/IMatchMap"
						},
						"type": "array",
						"description": "The maps which will be played. If match state is still ELECTION than this is not final."
					},
					"currentMap": {
						"type": "number",
						"format": "double",
						"description": "Index of the matchMaps array indicating the current map."
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "Send various events to this url (HTTP POST)"
					},
					"webhookHeaders": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"nullable": true,
						"description": "Additional headers that will be added to each webhook request"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"required": ["end", "match", "knife", "init"],
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"logs": {
						"items": {
							"$ref": "#/components/schemas/TLogUnion"
						},
						"type": "array"
					},
					"players": {
						"items": {
							"$ref": "#/components/schemas/IPlayer"
						},
						"type": "array"
					},
					"tmtSecret": {
						"type": "string",
						"description": "Access token to be used in the API."
					},
					"isStopped": {
						"type": "boolean",
						"description": "If match is finished or if the match was stopped/deleted this is true."
					},
					"serverPassword": {
						"type": "string",
						"description": "Server password, periodically fetched from game server"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"createdAt": {
						"type": "number",
						"format": "double",
						"description": "Creation date (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"lastSavedAt": {
						"type": "number",
						"format": "double",
						"description": "Last time the match was saved to disk (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					}
				},
				"required": [
					"id",
					"state",
					"mapPool",
					"teamA",
					"teamB",
					"electionSteps",
					"election",
					"gameServer",
					"logSecret",
					"parseIncomingLogs",
					"matchMaps",
					"currentMap",
					"webhookUrl",
					"webhookHeaders",
					"rconCommands",
					"canClinch",
					"matchEndAction",
					"logs",
					"players",
					"tmtSecret",
					"isStopped",
					"serverPassword",
					"createdAt",
					"mode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITeamCreateDto": {
				"description": "Team (create structure).",
				"properties": {
					"name": {
						"type": "string"
					},
					"passthrough": {
						"type": "string",
						"description": "Passthrough data to identify team in other systems.\nWill be present in every response/webhook."
					},
					"advantage": {
						"type": "number",
						"format": "double",
						"description": "Advantage in map wins, useful for double elimination tournament finals."
					},
					"playerSteamIds64": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Steam ids of players in \"Steam ID 64\" format. Will be forced into this team."
					}
				},
				"required": ["name"],
				"type": "object",
				"additionalProperties": false
			},
			"IMatchCreateDto": {
				"properties": {
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool.\nWorkshop maps are possible (contain numbers only).\nFriendly names for players when picking/banning maps can be added after a \"/\" delimiter and is advised for workshop maps.\n\nExample:\n```\n[\n   \"de_ancient\",\n   \"de_anubis/anubis\",\n   \"de_inferno\",\n   \"de_mirage\",\n   \"de_nuke\",\n   \"de_overpass\",\n   \"de_vertigo\",\n   \"3070923343/fy_pool_day\"\n]\n```"
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"electionSteps": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/IElectionStepAdd"
								},
								{
									"$ref": "#/components/schemas/IElectionStepSkip"
								}
							]
						},
						"type": "array"
					},
					"gameServer": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IGameServer"
							}
						],
						"nullable": true
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "Send various events to this url (HTTP POST)"
					},
					"webhookHeaders": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"nullable": true,
						"description": "Additional headers that will be added to each webhook request"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					}
				},
				"required": ["mapPool", "teamA", "teamB", "electionSteps", "gameServer"],
				"type": "object",
				"additionalProperties": false
			},
			"IMatchResponse": {
				"properties": {
					"id": {
						"type": "string",
						"description": "tmt2 identifier for this match"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchState"
					},
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool.\nWorkshop maps are possible (contain numbers only).\nFriendly names for players when picking/banning maps can be added after a \"/\" delimiter and is advised for workshop maps.\n\nExample:\n```\n[\n   \"de_ancient\",\n   \"de_anubis/anubis\",\n   \"de_inferno\",\n   \"de_mirage\",\n   \"de_nuke\",\n   \"de_overpass\",\n   \"de_vertigo\",\n   \"3070923343/fy_pool_day\"\n]\n```"
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team A for this match.\nTeam A will always be Team A in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team B for this match.\nTeam B will always be Team B in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"electionSteps": {
						"items": {
							"$ref": "#/components/schemas/IElectionStep"
						},
						"type": "array",
						"description": "List of election steps to determine the played map(s)."
					},
					"election": {
						"$ref": "#/components/schemas/IElection",
						"description": "Data for the election process."
					},
					"gameServer": {
						"$ref": "#/components/schemas/IGameServer"
					},
					"logSecret": {
						"type": "string",
						"description": "Log secret that is given as part of the url to the CS2 server as a log receiver (logaddress_add_http)."
					},
					"parseIncomingLogs": {
						"type": "boolean",
						"description": "Indicates if incoming logs from the CS2 server are parsed (otherwise they will be dropped without any action).\nWill be set to true if match is loaded from storage (after a short delay)."
					},
					"matchMaps": {
						"items": {
							"$ref": "#/components/schemas/IMatchMap"
						},
						"type": "array",
						"description": "The maps which will be played. If match state is still ELECTION than this is not final."
					},
					"currentMap": {
						"type": "number",
						"format": "double",
						"description": "Index of the matchMaps array indicating the current map."
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "Send various events to this url (HTTP POST)"
					},
					"webhookHeaders": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object",
						"nullable": true,
						"description": "Additional headers that will be added to each webhook request"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"required": ["end", "match", "knife", "init"],
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"logs": {
						"items": {
							"$ref": "#/components/schemas/TLogUnion"
						},
						"type": "array"
					},
					"players": {
						"items": {
							"$ref": "#/components/schemas/IPlayer"
						},
						"type": "array"
					},
					"tmtSecret": {
						"type": "string",
						"description": "Access token to be used in the API."
					},
					"isStopped": {
						"type": "boolean",
						"description": "If match is finished or if the match was stopped/deleted this is true."
					},
					"serverPassword": {
						"type": "string",
						"description": "Server password, periodically fetched from game server"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"createdAt": {
						"type": "number",
						"format": "double",
						"description": "Creation date (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"lastSavedAt": {
						"type": "number",
						"format": "double",
						"description": "Last time the match was saved to disk (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					},
					"isLive": {
						"type": "boolean",
						"description": "Match is currently supervised."
					}
				},
				"required": [
					"id",
					"state",
					"mapPool",
					"teamA",
					"teamB",
					"electionSteps",
					"election",
					"gameServer",
					"logSecret",
					"parseIncomingLogs",
					"matchMaps",
					"currentMap",
					"webhookUrl",
					"webhookHeaders",
					"rconCommands",
					"canClinch",
					"matchEndAction",
					"logs",
					"players",
					"tmtSecret",
					"isStopped",
					"serverPassword",
					"createdAt",
					"mode",
					"isLive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer_token": {
				"type": "http",
				"scheme": "bearer"
			},
			"bearer_token_optional": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "tmt2-backend",
		"version": "2.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Jens Forstmann"
		}
	},
	"paths": {
		"/api/stats/players": {
			"get": {
				"operationId": "GetPlayersStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlayerStats"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get global player statistics.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": []
			}
		},
		"/api/stats/players/match": {
			"get": {
				"operationId": "GetMatchPlayersStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlayerStats"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get the stats of all the players for a match.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats/matches": {
			"get": {
				"operationId": "GetMatchesStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMatchStats"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get global match statistics.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": []
			}
		},
		"/api/stats/match": {
			"get": {
				"operationId": "GetMatchStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMatchStats"
								}
							}
						}
					}
				},
				"description": "Get match stats for a specific match.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats/match/map": {
			"get": {
				"operationId": "GetMatchMapStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMatchMapStats"
								}
							}
						}
					}
				},
				"description": "Get match map stats for a specific match map.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "map",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats/player": {
			"get": {
				"operationId": "GetPlayerStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlayerStats"
								}
							}
						}
					}
				},
				"description": "Get the total stats of a player.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats/matches/player": {
			"get": {
				"operationId": "GetPlayerMatchesStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlayerStats"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get the stats of a player, divided by match.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats/team": {
			"get": {
				"operationId": "GetTeamPlayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all the players for a team.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/stats": {
			"delete": {
				"operationId": "DeleteStats",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": []
			}
		},
		"/api/matches": {
			"post": {
				"operationId": "CreateMatch",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IMatch"
										},
										{}
									]
								}
							}
						}
					}
				},
				"description": "Create and supervise a new match.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IMatchCreateDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllMatches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMatchResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all matches.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"description": "State filter",
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "Passthrough filter",
						"in": "query",
						"name": "passthrough",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "Get only stopped or not stopped matches.",
						"in": "query",
						"name": "isStopped",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Filter for only live (currently active) matches, or the opposite.",
						"in": "query",
						"name": "isLive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}